Workaround MPI_Bcast bug w.r.t. MPI_BOTTOM.
Some implementations have issues with  using MPI_BOTTOM in MPI_Bcast.
If that the case we can use an alternate implementation.

There is a macro in config.hpp named BOOST_MPI_BCAST_BOTTOM_WORKS_FINE
that can be used to control that. broadcast_test.cpp print a message
that point to that macro if suspicious behavior is detected.

LAM and Bull X MPI seems to have that problem.